添加weibo工具条 封装cell的weiboView 修改登陆认证时的逻辑 修改UIView (frameExtension)分类的数据类型(int->CGFloat)

+++ ADweibo/classes/Category/UIView+frameExtension.h
@@ -10,10 +10,10 @@
 
 @interface UIView (frameExtension)
 
-@property(nonatomic,assign)int X;
-@property(nonatomic,assign)int Y;
-@property(nonatomic,assign)int width;
-@property(nonatomic,assign)int height;
+@property(nonatomic,assign)CGFloat X;
+@property(nonatomic,assign)CGFloat Y;
+@property(nonatomic,assign)CGFloat width;
+@property(nonatomic,assign)CGFloat height;
 
 
 @end
